fix #119897

+++ cache2k-core/src/main/java/org/cache2k/core/ClockProPlusEviction.java
@@ -194,7 +194,7 @@ public class ClockProPlusEviction extends AbstractEviction {
   }
 
   public long getSize() {
-    return hotSize + coldSize;
+    return (long)hotSize + (long)coldSize;
   }
 
   @Override
